pyproject.toml
[tool.poetry]
name = "project-folder"
version = "0.1.0"
description = ""
authors = ["Eze Onyekachukwu"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.3"
django-environ = "^0.11.2"
psycopg2 = "^2.9.9"
djangorestframework = "^3.15.1"
django-filter = "^24.2"
pillow = "^10.2.0"
celery = "^5.3.6"
redis = "^5.0.3"
django-celery-results = "^2.5.1"
django-celery-beat = "^2.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
